remove the not used main template/controller/route
+ the 'main' route, nested under 'app', which contains the tabs, can
also be combined with the sidemenu.

+ the template contents of main.html are thus moved into the sidemenu
template, where the named nav-view is replaced with the contents of the
main template

+ as the controller did not have any actions, it is deleted without
replacement, and the reference to it in index.html is removed

+ the (abstract) state app.main no longer exists, thus moving the 'tab'
states to app.tab1, and app.tab2 repsectively.

+ this reduces the number of files, simplifies the urls for the routes,
and combines all the view html code for the navigation skeleton in the
sidemenu.html file